[REFACTOR] Reimplement download functionality

+++ spider_pixiv.py
@@ -1,5 +1,5 @@
-import json
 import time
+import random
 import requests
 import os
 import re
@@ -11,6 +11,8 @@ from selenium.webdriver.support import expected_conditions as EC
 from selenium.common.exceptions import TimeoutException
 from selenium.common.exceptions import NoSuchElementException
 
+
+
 # 预定义不允许的字符
 INVALID_CHARS = r'[<>:"/\\|?*]'
 
@@ -20,6 +22,10 @@ base_file = os.path.abspath(__file__)
 # 获取文件的文件夹信息
 base_dir = os.path.dirname(base_file)
 
+
+# ? 需要查清楚是连接出现了问题，还是 selenuim 执行 js 出现了问题
+
+
 class downloadImg:
     def __init__(self, save_dir, illust_name, illust_num, illust_url):
         self.save_dir = save_dir
@@ -28,7 +34,6 @@ class downloadImg:
         self.illust_url = illust_url
 
     def download_illust(self):
-
         _, illust_ex = os.path.splitext(self.illust_url)
 
         img_name = f"{self.illust_name}-{self.illust_num}{illust_ex}"
@@ -46,13 +51,17 @@ class downloadImg:
                 img_response = session.get(self.illust_url)
 
                 if img_response.status_code == 200:
+                    if os.path.exists(save_path):
+                        print(f"{img_name} 已存在")
+                        return
+
                     with open(save_path, "wb") as img_file:
                         img_file.write(img_response.content)
 
                     print(f"{img_name} 下载完成")
                     return
 
-                elif img_response == 429:
+                elif img_response.status_code == 429:
                     retry_count += 1
 
                     wait_time = 5 * retry_count  # 等待时间
@@ -102,8 +111,6 @@ class createDir:
 
     #  创建作品文件夹
     def createIllustDir(self):
-
-
         # 确保 author_dir_path 不为 None
         if not self.author_dir_path:
             print("请先创建 author 目录")
@@ -119,7 +126,7 @@ class createDir:
 
     def return_save_path(self):
         return self.illust_dir_path
-    
+
     def return_illust_name(self):
         return self.illust_name
 
@@ -146,6 +153,7 @@ class getOriginal:
         for illustsID in self.illusts_id:
             url = f"https://www.pixiv.net/artworks/{illustsID}"
 
+
             self.driver.get(url)
 
             # 作者名称
@@ -154,18 +162,43 @@ class getOriginal:
                 '//*[@id="root"]/div[2]/div/div[3]/div/div/div[1]/aside/section[1]/h2/div/div/a/div',
             ).text
 
-            # 作品名称
-            self.illust_text = self.driver.find_element(
-                By.XPATH,
-                '//*[@id="root"]/div[2]/div/div[3]/div/div/div[1]/main/section/div[1]/div/figcaption/div/div/h1',
-            ).text
+            try:
+                # 作品名称
+                self.illust_text = self.driver.find_element(
+                    By.XPATH,
+                    '//*[@id="root"]/div[2]/div/div[3]/div/div/div[1]/main/section/div[1]/div/figcaption/div/div/h1',
+                ).text
+
+            except NoSuchElementException:
+                self.illust_text = "无题"
+
+            if self.illust_text == "无题":
+
+                try:
+                    illust_date = WebDriverWait(self.driver, 10).until(
+                        EC.visibility_of_element_located(
+                            By.XPATH,
+                            '//*[@id="root"]/div[2]/div/div[3]/div/div/div[1]/main/section/div[1]/div/figcaption/div/div/div[4]/time',
+                        ).text
+                    )
+
+
+                    match = re.search(r"(\d{4})年(\d{1,2})月(\d{1,2})日", illust_date)
+                    if match:
+                        year, month, day = match.groups()
+                        # 格式化成 'YYYY-MM-DD' 的格式
+                        date_str = f"{year}-{int(month):02d}-{int(day):02d}"
+
+                        self.illust_text = self.illust_text + "-" + date_str
+                except Exception as e:
+                    print(f"无题作品发生了异常: {e}")
 
             new_dir = createDir(
                 author_name=self.author_text, illust_name=self.illust_text
             )
-            
+
             new_dir.createResultDir()
-            
+
             self.illust_text = new_dir.return_illust_name()
 
             if self.author_dir_flag:
@@ -173,7 +206,6 @@ class getOriginal:
             else:
                 new_dir.createAuthorDir()
                 self.author_dir_flag = True
-
 
             new_dir.createIllustDir()
 
@@ -193,102 +225,66 @@ class getOriginal:
                     '//*[@id="root"]/div[2]/div/div[3]/div/div/div[1]/main/section/div[1]/div/figure/div/div[1]/div/div/div/div/div/span',
                 )
 
-
             except NoSuchElementException:
                 page_element = False
 
             except Exception as e:
-                print(f"发生异常:{e}")
+                print(f"页面多图标签发生异常: {e}")
 
             if page_element:
                 print("确认为多图")
+                
                 self.flag = True
+                self.get_id = illustsID
+                
+                image_info = page_element.text # 页面多图标签的文字
+                self.image_num = int(image_info.split('/')[-1])  # 以斜杠为分割符
             else:
                 print("确认为单图")
+                self.get_id = illustsID
+                self.image_num = 1
+                
 
             self.getOriginalUrl()
 
 
     def getOriginalUrl(self):
-
-        # 确认为多图
-        if self.flag:
-
-            # 等待 "查看全部" 按钮加载，并点击它
-            try:
-                view_all_button = WebDriverWait(self.driver, 10).until(
-                    EC.element_to_be_clickable(
-                        (
-                            By.XPATH,
-                            '//*[@id="root"]/div[2]/div/div[3]/div/div/div[1]/main/section/div[1]/div/div[4]/div/div[2]/button/div[2]',
-                        )
-                    )
-                )
-                view_all_button.click()
-
-            except Exception as e:
-                print(f"发生异常: {e}")
-
-
-            # 应对懒加载，下拉页面
-
-            self.driver.execute_script("window.scrollTo(0,10000);")
-
-            time.sleep(1)
-
-            self.driver.execute_script("window.scrollTo(0,10000);")
-
-            time.sleep(1)
-
-            self.driver.execute_script("window.scrollTo(0,10000);")
-
-            try:
-                a_tags = WebDriverWait(self.driver, 10).until(
-                    EC.presence_of_all_elements_located(
-                        (By.CSS_SELECTOR, "a.sc-1qpw8k9-3.ilIMcK")
-                    )
-                )
-
-            except Exception as e:
-                print(f"获取图片出现异常: {e}")
-
+        
+            
+        url = f'https://www.pixiv.net/ajax/illust/{self.get_id}'
+            
+            
+        response = self.session.get(url)
+        if response.status_code == 200:
 
             self.illust_num = 1
-            
-            # 提取每个 <a> 标签的 href 属性
-            for a_tag in a_tags:
-                img_url = a_tag.get_attribute("href")
                 
-                download_img = downloadImg(self.save_dir,self.illust_text,self.illust_num,img_url)
-                
-                download_img.download_illust()
-                
-                self.illust_num += 1
-
-
-        else:
-
-            try:
-                a_tags = WebDriverWait(self.driver, 10).until(
-                    EC.presence_of_all_elements_located(
-                        (By.CSS_SELECTOR, "a.sc-1qpw8k9-3.ilIMcK")
-                    )
+            data = response.json()
+            original_url = data.get("body",{}).get("urls",{}).get("original")
+                    
+            url_prefix , url_suffix = original_url.split('_p')
+                    
+            _ , ex_name = os.path.splitext(url_suffix)
+                    
+            for i in range(self.image_num):
+                img_url = f'{url_prefix}_p{i}{ex_name}'
+                      
+                download_img = downloadImg(
+                self.save_dir, self.illust_text, self.illust_num, img_url
                 )
 
-            except Exception as e:
-                print(f"获取图片出现异常: {e}")
-
-            # 提取每个 <a> 标签的 href 属性
-            for a_tag in a_tags:
-                img_url = a_tag.get_attribute("href")
-                
-                download_img = downloadImg(self.save_dir,self.illust_text,self.illust_num,img_url)
-                
                 download_img.download_illust()
 
                 self.illust_num += 1
 
-        time.sleep(5)
+                    
+            else:
+                print(f"原图url请求失败，状态码: {response.status_code}")
+
+
+
+        time.sleep(3)
+
 
 class getAuthorProfile:
     def __init__(self, driver, authorID):
@@ -308,6 +304,7 @@ class getAuthorProfile:
         # 设置User-Agent和其他必要的头信息
         headers = {
             "User-Agent": user_agent,
+            # 'Referer': 'https://www.pixiv.net/',
             "Referer": "https://www.pixiv.net/tags/%E7%BA%B1%E9%9B%BE",
         }
         session.headers.update(headers)
@@ -331,6 +328,7 @@ class getAuthorProfile:
 
         return illusts_id, session
 
+
 class autoLogin:
     def __init__(self, driver, email, password):
         self.driver = driver
@@ -340,22 +338,37 @@ class autoLogin:
     def auto_login(self):
         self.driver.get("https://accounts.pixiv.net/login")
 
+
         self.driver.implicitly_wait(10)
 
-        self.driver.find_element(
+
+
+        input_email = self.driver.find_element(
             By.XPATH,
             '//*[@id="app-mount-point"]/div/div/div[4]/div[1]/div[2]/div/div/div/form/fieldset[1]/label/input',
-        ).send_keys(self.email)
+        )
 
-        self.driver.find_element(
+        for char in self.email:
+            time.sleep(random.uniform(0.1, 0.5))
+            input_email.send_keys(char)
+
+        input_password = self.driver.find_element(
             By.XPATH,
             '//*[@id="app-mount-point"]/div/div/div[4]/div[1]/div[2]/div/div/div/form/fieldset[2]/label/input',
-        ).send_keys(self.password)
+        )
+
+        for passwd_char in self.password:
+            time.sleep(random.uniform(0.1, 0.5))
+            input_password.send_keys(passwd_char)
+
+        time.sleep(random.uniform(1, 3))
 
         self.driver.find_element(
             By.XPATH,
             '//*[@id="app-mount-point"]/div/div/div[4]/div[1]/div[2]/div/div/div/form/button',
         ).click()
+
+
 
         try:
             WebDriverWait(self.driver, 20).until(
@@ -367,18 +380,24 @@ class autoLogin:
                 )  # 找到用户的头像
             )
 
+            # //*[@id="root"]/div[2]/div/div[2]/div[1]/div[1]/div/div[3]/div[1]/div[5]/div/button/pixiv-icon//svg
+
             print("登录成功")
         except TimeoutException:
             print("登录超时，可能登录失败")
 
         return self.driver
 
+
 class initializeConfig:
     def __init__(self, user_agent):
+        # self.proxy = proxy
         self.user_agent = user_agent
 
     def setup_WebDriver(self):
         chrome_options = Options()
+        # chrome_options.add_argument(f"--proxy-server={self.proxy}")
+
         chrome_options.add_argument(
             "--allow-running-insecure-content"
         )  # 允许https中加载http
@@ -388,7 +407,10 @@ class initializeConfig:
         chrome_options.add_experimental_option(
             "excludeSwitches", ["enable-automation"]
         )  # 设置绕过selenium 检测
+
         chrome_options.add_argument("--ignore-ssl-errors")  # 忽略SSL错误
+        # chrome_options.add_argument('--allow-insecure-localhost')  # 允许访问不安全的 localhost（自签名证书）
+        # chrome_options.add_argument("--headless")  无头模式，即不显示浏览器
         chrome_options.add_argument("--disable-gpu")
         chrome_options.add_argument("--disable-dev-shm-usage")  # 禁用共享内存
 
@@ -400,23 +422,27 @@ class initializeConfig:
         return webdriver.Chrome(options=chrome_options)
 
 
+
 if __name__ == "__main__":
     # 记录程序开始时间
     start_time = time.time()
+
 
     email = "SET YOUR EMAIL"
     password = "SET YOUR PASSWD"
 
     user_agent = "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/58.0.3029.110 Safari/537.36"
 
-    authorID = "39750820"  # 测试作者2
+
+    authorID = input("输入作者的PID: ")
+
 
     init_config = initializeConfig(user_agent)
 
     driver = init_config.setup_WebDriver()
 
     login = autoLogin(driver, email, password)
-    
+
     login.auto_login()
 
     profile = getAuthorProfile(driver, authorID)
@@ -426,6 +452,7 @@ if __name__ == "__main__":
     original = getOriginal(author_illusts, session, driver)
 
     original.confirmImgNum()
+
 
     # 记录程序结束时间
     end_time = time.time()
